ADD : revised OSB format. Allows to use large map files directly from disk with low memory footprint. FIX : Native render doesn't render anything at high lat/long values (Closes #1135) FIX : Expand tables in the properties dock to the available width by default (Closes #1136)
git-svn-id: http://svn.openstreetmap.org/applications/editors/merkaartor@10406 b9d5c4c9-76e1-0310-9c85-f3177eceb1e4

+++ Map/ImportOSM.cpp
@@ -135,7 +135,7 @@ void OSMHandler::parseWay(const QXmlAttributes& atts)
 		else if (R->lastUpdated() != MapFeature::UserResolved)
 		{
 			while (R->size())
-				R->remove(0);
+				R->remove((unsigned int)0);
 			NewFeature = true;
 			if (R->lastUpdated() == MapFeature::NotYetDownloaded)
 				R->setLastUpdated(MapFeature::OSMServer);
@@ -196,7 +196,7 @@ void OSMHandler::parseRelation(const QXmlAttributes& atts)
 		else if (R->lastUpdated() != MapFeature::UserResolved)
 		{
 			while (R->size())
-				R->remove(0);
+				R->remove((unsigned int)0);
 			NewFeature = false;
 			if (R->lastUpdated() == MapFeature::NotYetDownloaded)
 				R->setLastUpdated(MapFeature::OSMServer);
